Update right-click enable
I just gave permission the right-click , it was kinda a experiment it seems good but in reality annoying. Also it does not good for mobile version if someone try to copy text the alert pop up again. So kinda a problem. I converted as a comment the right-click code maybe in the future I can work on.