Disable logo hover-active for mobile
I tried and didn't like the feature in mobile version so I decided to disable for that, but still working on web version and it's so good I loved that. Also I added a code for background image changes in local files instead of committing directly, so it is for test purposes.