Increase logo height
Normally, it was good looking for "hover" function when someone hovers over it, to change its rotation etc. However, I found a glitch that when someone hovers over the border of the logo, not the center, the logo becomes much smaller and the hover function does not work because technically it is not on the logo. So I increased the height of the logo and now it fits and looks good.